requirements.txt
fastapi 
uvicorn 
httpx 
supabase 
python-dotenv 
pydantic
requests